requirement.txt
Django==3.0.5
#django-auth-ldap==2.4.0
django-easy-audit==1.3.0
django-notifications-hq==1.6.0
django-environ==0.4.5
django-rq==2.4.1
redis==3.5.3
configparser==5.0.2
xlsxwriter==1.4.0
psycopg2-binary==2.8.6
cx_Oracle==8.1.0
lxml==4.6.3
alive_progress==1.6.2
tqdm==4.60.0
telegram-send==0.25